test(huggingchat): isolate stream boundary state
